refactor: move spriteTypes into entiry
