+++ frontend/src/game/entity/Placeholder/states.ts
@@ -1,4 +1,4 @@
-import { StateConfig } from '../types';
+import { StateConfig } from '..';
 
 enum PlaceholderState {
   Idle,
