+++ frontend/src/game/entity/types/index.ts
@@ -2,7 +2,9 @@ import Direction from './direction';
 import StateConfig from './stateConfig';
 import EntityClass from './entityClass';
 import EntityId from './entityIds';
+import SpriteAnimation from './spriteAnimation';
+import SpriteConfig from './spriteConfig';
 
 export {
-  Direction, StateConfig, EntityId, EntityClass,
+  Direction, StateConfig, EntityId, EntityClass, SpriteAnimation, SpriteConfig,
 };
