+++ frontend/src/game/entity/types/spriteConfig.ts
