+++ frontend/src/game/sprites/character/walk/index.ts
@@ -1,5 +1,5 @@
 import sprite from './index.png';
-import SpriteConfig from '../../spriteConfig';
+import { SpriteConfig } from '../../../entity';
 import { Point } from '../../../shapes';
 
 const sc:SpriteConfig = {
